istests return False if not Qobj
If the object that is tested with the functions in istests.py is not
a Qobj the functions now return False, whereas previously they would
have raised an exception.

In order to do that I had to put the istests into Qobj in order to
avoid cyclical import. In istests.py the tests are imported for
backwards compatability.