Added new logging and settings support, started refactoring debugs.
This change enables the use of Python standard logging, and backs
it by a new qutip.settings entry. A few instances of classic
'if debug: print' style logging have been refactored.
The new qutip.settings entry is based on configobj, with a fallback
to classic qutiprc parsing if that library is not available.
Errors in the new functionality introduced by this commit are also
logged, and can be accessed by setting up logging.basicConfig
*before* importing QuTiP.