Send numpy array to cy_td_ode_rhs
This tries to fix deprecation warnings in the travis tests. The function
_td_ode_rhs_super in mesolve.py previously sent an "array-slice" to the
cython function cy_td_ode_rhs. This appears to cause a deprecation
warning "DeprecationWarning: using a non-integer number instead of an
integer will result in an error in the future", which causes travis
tests to fail. This is now hopefully fixed by sending a numpy array in
stead of the "array-slice".