Fix use of round in rhs_generate.py
This commit attempts to avoid deprecation warnings in travis tests:
"DeprecationWarning: using a non-integer number instead of an
    integer will result in an error in the future"
The cause of the deprecation warning seems to be in rhs_generate.py, where
an array index is returned using the round() function, which returns a
float, not an integer. The call to round() is now wrapped in a call to
int().