Update hermitian tests
The tests seems to fail for the Hermitian mkl_spsolve tests under Py2.7.  However, I can not reproduce this outside of the testing environment.  This update makes the tests a bit better behaved.  I have run this under Py2.7 (on mac and win) over a million times, and the solution error is never above 1e-15.