Fix bug in _ode_rho_func_td*
The functions _ode_rho_func_td (never actually used but useful for
debugging purposes) and _ode_rho_func_td_with_state had a bug in that
the call to L_func(t, args) / L_func(t, rho, args) [assumed to return
Qobj's] was not converted to sparse matrix.

This functionality is outdated according to the manual, but still good
to have functions that work.