Fix an issue with expectation values computation in fmmesolve
When writing the density matrix in the original computational basis,
from the Floquet basis, there was an error in the transformation used,
actually doing the inverse of what it should. This was resulting in bad
computed expectation values.

This commit also adds a couple of tests for the simulation of
dissipative systems using Floquet formalism.