Fast Husimi Q using precomputation
A lot of the cost of the Husimi Q function does not depend on the state.
If it is called many times (e.g. for a density matrix or when doing
multiple states), this can be used for a speedup. I wrote a function to
precompute everything that does not depend on the state and store it as
a 3d array. The Q function itself is then a dot product between the last
axis of the precomputed array and a pure state.

Performance:
  - 3x for N=10
  - 4x for N=20
  - 7x for N=50
Everything tested with 401 points in x and y, respectively.  For density
matrices, precomputing is already paid for by the first call. Because
the memory cost can be large, I still leave the decision to the user.

The third if statement in _qfunc_pure can be avoided at a small
performance cost in either the polyval or np.dot variant. This seemed
the best compromise between performance and nice code.

Co-authored-by: Jake Lishman <jake@binhbar.com>