Convert test_brmesolve_td.py to pytest
The `test_simple_qubit_system` tests here could possible be dropped;
except for the use of the string to specify the spectrum function and
the unnecessary `H = [[H, '1']]`, this tests all the same paths as
`test_brmesolve.test_simple_qubit_system`---I ported to `pytest` mostly
just be copy-pasting the code.  Both of those changes are tested
elsewhere in this file, so we wouldn't be losing anything.