Convert test_control_pulseoptim.py to pytest
Converted the setup/tear down method of temporary file handling to one
making better use of pytest's fixtures.  This also swaps the temporary
files to the standard library `tempfile`, which ensures that we're not
(even momentarily) messing around with the user's home directory, but
that we use the OS-supplied temporary folder.  All creation and cleanup
is then just handled automatically with context managers, whether
explicitly when we use temporary files, or implicitly by pytest.

The massive complexity of the `load_parameters_from_file` test would
suggest that that particular module isn't organised as well as it could
be.  It was rather difficult getting the test to work at all, let alone
actually using it as an end-user.