Convert test_correlation.py to pytest
The custom 2D integration in the two-level system had a bug caused by
using `sum()` on a 2D numpy array: `sum` works over iterables, so this
is functionally equivalent to `np.sum(x, axis=0)` and _not_ `np.sum(x)`
which is what was intended for the trapezium rule to have been
implemented correctly.  This left `s3` (the centre of the region) still
being an array, while `s1` and `s2` (the corners and edges respectively)
were correctly scalars.  The broadcasting rules in the line
    sum(s1 + 2*s2 + 4*s3)
then effectively meant that `s1` and `s2` were multiplied by
`s3.shape[0]`.

I removed the custom 2D integration, and just used multiple calls to
`np.trapz` to ensure accuracy in the future.  This was already being
done in the three-level system.  I updated the expected values of `g2_0`
to match the correct integration.