Add in_temporary_directory fixture
This fixture is neater to use than the previous implicit `tmpdir`
prodived by `pytest`.  The marked tests run in a temporary directory
that is automatically cleaned up at the expiration of the fixture.  This
invalidates the need for tests to ensure they are in a nice directory,
and means that in the case of any "by-product" files being dumped
(in particular the `QobjEvo` class has a habit of dropping ".pyx" files
all over the place) then they'll get cleaned up even if the user kills
the tests.

This is used as a wrapper fixture only for its setup and tear down
code; nothing is actually passed to the function.