Deprecate qutip.graph functions
The `qutip.graph` functions are pure mathematical functions which are
beyond the purview of the package.  Most of them have already been
merged into `scipy.sparse.csgraph`, so we now move to using those and
mark the functions as deprecated.

Only two functions are actually used, and they're not even currently
tested.  `weighted_bipartite_matching` is not present in `scipy`, but is
still marked as deprecated because we can move it into being an
internal-only function, or simply remove the code that depends on it
since those paths are not tested.

This will allow the removal of a good amount of Cython code, and
simplify maintenance.

See #1203 for further discussion.