Convert test_heom_solver.py to pytest
Separate out the different test cases into parametrised tests.  At
present the expected value for each test is the same and re-calculated
on each test run.  This isn't much of an issue because the calculation
time is short (<0.1s on my machine), but if it becomes an issue, it's a
fairly simple matter of caching the result.  I just didn't want to
overengineer a simple conversion.

The test that `hsolver.configure` correctly handles setting and
unsetting the stats is moved into a separated test to uncouple the
logic.