4.5.1v2 (#1215)
* Patch4.5 fix1169 (#1191)

* name collision fix for parallel propagator (#1169)

* name collision fix for parallel propagator

* pep8

* update setup release to micro 4.5.1

Co-authored-by: Eric Giguère <eric.giguere@calculquebec.ca>

* Boxi Li added to authors and admins (#1193)

* Issues about errors (#1199)

* small fixes

* split PR

* error in cqobjevo

* split PR

* omp also

* Activate Python 3.8 builds on Linux. (#1170)

* Fix clebsch function not working with non-int inputs (#1177)

* Add files via upload

* tidying up

* Update test_clebsch.py

* Update test_clebsch.py

* Update test_clebsch.py

* Update utilities.py

* Delete test_clebsch.py

* Deprecate qutip.graph functions (#1206)

* Fix pep8 in graph.py

* Deprecate qutip.graph functions

The `qutip.graph` functions are pure mathematical functions which are
beyond the purview of the package.  Most of them have already been
merged into `scipy.sparse.csgraph`, so we now move to using those and
mark the functions as deprecated.

Only two functions are actually used, and they're not even currently
tested.  `weighted_bipartite_matching` is not present in `scipy`, but is
still marked as deprecated because we can move it into being an
internal-only function, or simply remove the code that depends on it
since those paths are not tested.

This will allow the removal of a good amount of Cython code, and
simplify maintenance.

See #1203 for further discussion.

* fix typo in QobjEvo._shift (#1233)

* add test for half integer for clebsch (#1218)

* Prefix variable names with _ to clear Travis CI warnings (#1213)

* fix conflict

* Fix spin husimi/wigner functions (#1195)

* Fix bug in spin_q_function for half-integer spins

Previous code only allowed integer spins. Also normalize the Q function with 1/pi pre-factor.

* Fix bug in spin_wigner for half-integer spins

Previous code only allowed integer spins.

* Fix PEP8 style

Co-authored-by: Nathan Shammah <nathan.shammah@gmail.com>
Co-authored-by: Alexander Pitchford <alex.pitchford@gmail.com>
Co-authored-by: Simon Cross <hodgestar+github@gmail.com>
Co-authored-by: Thomas Walker <42548316+taw181@users.noreply.github.com>
Co-authored-by: Jake Lishman <jakelishman@gmail.com>
Co-authored-by: Ivan Carvalho <ivan.ivancps.cn@gmail.com>
Co-authored-by: maij <maij@users.noreply.github.com>