V4.5.1v3 (#1241)
* Udated tests to 3.8 (#1214)

* moved tests to 3.8

* safer mcsolve test for pool

* typos in .travis

* force fork on mac

* pep8

* typos

* remove 3.6 and shorted 3.7

* merge tests

* Fix typo in test name

* set_context to get_context

* fix dtype for randint (#1219)

* Fix error in definition of m_ops for heterodyne detection in smesolve() (#1237)

* Error in definition of m_ops for heterodyne detection in smesolve()

There was an error in line 613 of qutip/stochastic.py as the measurement operators for heterodyne detection in the implemented stochastic master equation should be `c + c.dag()` and `1j (c - c.dag())` as explicitely presented in equation (5.64) and the following 2 paragraphs in https://arxiv.org/abs/1710.09523. The same measurement operators can also be drawn from equation (4.108) of Milburn & Wiseman _Quantum Measurement and Control_.

* Fix same m_ops error for _positive_map

Fixed the same error in defining the m_ops, now present in _positive_map which is used for the 'rouchon' solver.

* Fix to correct sign convention

Added a minus sign to be consistent with the implementation of the sops just below and also with the ssesolve() function which had the correct implementation of the m_ops initially, i.e. before this pull request. See line 737.

Co-authored-by: eliegenois <55160591+eliegenois@users.noreply.github.com>