Add dispatcher prototype
The dispatcher will be used for handling multiple-dispatch over
different data type inputs and outputs.  We try to ensure a lookup time
that is constant in the number of data types eligible for dispatch, and
linear in the number of dispatched parameters.

As a base, the dispatch is done using a hash table lookup on a tuple of
the provided classes, and dispatches only if an exact match for
specialiisation is found.  Specialisations are provided at runtime
by a Python decorator, so that the dispatched method does not itself
need to be aware of every possible specialisation, and more
specialisations can be added later by other packages if suitable.

To do the dispatch at call time, we first have to emulate the Python
binding process to match up all supplied arguments with the inputs we
are expecting.  A pure-Python implementation could be made simply using
inspect.Signature.bind, but this is very slow (~10µs).  Instead, we
provide a Cythonised binder which constructs helpful datatypes at
intialisation of a dispatcher function, which cuts dispatch time down to
~500ns for a positional-only call. Interpreting the `kwargs` dictionary
will incur a time penalty in Cython, on the order of 100ns per keyword
passed, so we always attempt to support positional arguments as the
fast path.