Add copy, nnz and sort_indices to CSR
For copy we heavily use memcpy for the most efficient copying.  We do
not copy any scipy.sparse.csr_matrix instance that may be backing the
original's data, since this is just a waste of time unless the user
requests it.

nnz would be better implemented as a pre-processor macro, but this is
unlikely to ever be a bottleneck anywhere, and Cython doesn't support
that kind of metaprogramming anyway on Cython types.

sort_indices is directly taken from the old .core.cy.sparse_routines
version.