Remove unnecessary memory imports
These functions _may_ call the Python memory allocator, and so cannot be
used without hold the GIL.  Instead we use malloc/calloc/free to ensure
that we're managing the memory ourselves.  This is always safe in these
cases, because the memory is only allocated to be temporary space
within a single scope.