Remove out parameter from matmul_csr
At present, no other operation allows this output parameter, and it is
mostly just asking for segfaults; the user typically would not know how
much space to allocate for the data and col_index arrays.

This could be re-added in the future with entry points from C code only,
if there are use cases where the amount of memory to be allocated is
known, or the underlying structures are changed to use C++ vector and
automatically expand if required - this is a non-trivial change,
however, if the type has active numpy array views onto the data.

The out parameter of matmul_vector_csr is not similar, because in that
case, it is not an operation between two data-layer objects, but between
a data-layer object and an ndarray, outputting an ndarray.