Support scipy >= 1.5 in fast_csr_matrix matmul
Scipy 1.5 changed the methods of performing the matrix multiplication to
avoid copying the indptr matrix multiple times.  This is actually not an
issue at all for us, because we only allow int32 indices (so if the
indices were to upcast, we just throw and exception), but the underlying
private Cython call structure also changed, and our fast_csr_matrix was
a direct clone.

We add try/catch blocks to switch on the correct types (the overhead
from a failed lookup is trivial compared to the matrix multiplication
time), because we have to support more than just the most recent version
of scipy.  We could look up the version of scipy installed, but the
"try it and see" approach is more typical Python.

This problem should be eliminated by the move to the new data types, as
we will stop relying on private scipy code.

See: commit scipy/scipy@53fac7a