Fix brtools zheevr test
The new test explicitly checks the eigenvalue equation for each
generated eigenvalue/vector pair, and tests (via the determinant) that
all eigenvectors are linearly independent, as they must be since they
are eigenvectors of a Hermitian matrix.

The previous test relied on the implementation of zheevr producing
_exactly_ the same result as scipy.linalg.eigh, which was very fragile,
and indeed broken by scipy 1.5.