Add isherm function for CSR
This is _not_ a direct port of the original spmath.zcsr_isherm.  We get
a good (2-3x) speedup by replacing the expensive call to "abs" with a
custom comparator which avoids the square root.  We don't actually need
the absolute value, so this is a simple change.

The speedup is still present when comparing non-Hermitian matrices, but
the only real benefit here is due to the improved access times, because
the short-circuited return means we will likely have made very few
comparisons.

Error handling readability and early return logic improved by
try/finally instead of for/else, break/continue.