Fix quadratic complexity in zcsr_inner
The bra--ket inner product had an unnecessary loop when looking up the
row in the ket, where we already knew which element we needed to lookup.

This function does not check the shapes of the inputs are valid and can
therefore segfault on improper input, but it's a low-level Cython
function, so that's expected.  These segfaults already existed, although
previously only in the ket--ket product, not the bra--ket product.