Fix zcsr_proj for bras with unsorted indices
Due to the CSR format, it is possible for a bra to have unsorted indices
(though really you'd try very hard to contrive a situation where this
was the case), but it is not possible for a ket.

Sorting is (worst case) O(nnz lg(nnz)), whereas the projection itself is
O(nnz**2), so there's very little runtime penalty for doing things the
right way.  Also, projection of a bra is pretty uncommon - most users
will have a ket, and this path wasn't broken.