Add CSR projection
Unlike the previous implementation in spmath.pyx, we do not take an
argument specifying whether we were passed a bra or a ket, we simply
read it out of the shape of the matrix, which simplifies the calling
structure.

This also ensures that indices are sorted for bra projection, just as PR
gh-1307 did for the original implementation.