Add inner product for CSR matrices
Unlike the previous implementation in spmath.pyx, we do not generally
require the user to pass a flag telling us whether the left side is a
bra or a ket, but read it out of the shape.  The sole exception is if
the shape of both matrices is (1, 1), in which case it could be either,
and so there is the special flag `scalar_is_ket`, which is only read in
the scalar case.

This also does not have the quadratic scaling for the bra--ket product
that the previous version did (fixed in gh-1306), but it does _not_ do
the shape checking which was introduced in that same PR.  In the data
layer model, we will define that callers of the data layer functions are
responsible for type-checking, so that in low-level Cython code we can
check only once at the start, then never need to again.  Qobj will be
updated in the new version to ensure that this is the case.