Add reshape for CSR matrices
This method removes the need for the COO/CSR conversion, saving a lot of
code.  It is a lot more consistent in time now, so
`reshape_csr(x, a, b)` will generally take the same amount of time (give
or take) as `reshape_csr(x, b, a)` unless the difference is really
really big.  Previously, it was much slower when b > a.  We're now
reliably a bit faster than the fast case before, but there's not much in
it for large matrices (for small ones our allocation is much faster).