Speed up reshape_csr
Remove the branching within the for-loop, since branch prediction would
not be especially effective on it.  It's faster just to do the quotient
and remainder, and then cumsum the row indices at the end, so there's no
branching at any point once the calculation starts properly.

This really makes things about ~2x faster for very sparse matrices being
reshaped to have many columns.

It's slightly faster to use libc.stdlib.div for very dense matrices
(~10% speedup for a full 100x100 matrix) as opposed to the standard //
and %.