Add <a|op|b> (inner_op/mat_elem) for CSR
This was previously called zcsr_mat_elem in spmatfuncs, but since the
data layer is more concerned with mathematical operations, it gets a
rename to inner_op, since it is the same inner product as inner, but
with an operator applied in the middle.

The order of arguments is also changed so that the "mathematical
function" call looks more like what its name suggests, i.e.
    inner_op(a, op, b) == <a|op|b>
Like inner, we use the shape parameter of `a` (aka `left`) to determine
whether we have been passed a bra or a ket, and supply an optional
keyword argument `scalar_is_ket` to determine behaviour for 1D systems.

The algorithms used for inner_op are new, and not ported from
zcsr_mat_elem.  They are noticeably faster for the ket--ket form---
between 10x faster for very small matrices to about ~1.5-3x faster at
larger ones, dependent on the sparsity.  For very large (1000x1000),
dense (density > 0.1) operators, the new form can be up to about ~5%
slower for ket--ket, but this is a worthwhile trade-off, I believe.

For bra--ket the algorithm is hugely improved, and works regardless of
whether the indices are sorted (the current implementation cannot handle
unsorted bra indices).  It is now effectively the same as the ket--ket
algorithm (differences now mostly due to caching/branch prediction/loop
unrolling intricacies, I think).  The old bra--ket algorithm performed
2-20x slower than the ket--ket.