Estimate shape with warning if not passed to CSR
The shape cannot be accurately inferred from three passed arrays, so it
should not really be valid to omit the shape at all.  Scipy and old
QuTiP both allow it, though, so we maintain it but emit a warning.
Since we have emitted a warning, it seems fair if we take a long time
trying to guess the correct shape, hence we no longer just assume that
it's square, but instead make an effort to get it as close as we can.