Fix null dereference or double free in CSR
If an exception occurred while calling or executing CSR.__init__, the
object CSR._scipy would be set to None when CSR.__dealloc__ was called,
triggering the deallocation procedure.  If the exception occurred before
CSR.data, CSR.col_index and CSR.row_index were set, then accessing
&self.data[0] would be a null-pointer dereference; self.data is a
`struct __pyx_memoryview` (or similar), and the pointer we want is the
`data` member of that struct.  Cython does not let us access that
directly, though, hence the funny dereference/reference.  Previously,
initializedcheck was set to False, so Cython would not detect it, and
we would enter undefined behaviour in the compiler.

To fixed the null dereference, we turn on initializedcheck in
__dealloc__, and use try/except/pass to safely jump if the memoryview is
not initialized.  This likely should never occur anyway, becuase of the
second change.  Further, Cython memoryviews through an error when you
try to initialise them with a null pointer, so we know for sure that any
contained pointer is not null, meaning that if not AttributeError is
thrown, the subsequent reference and call to free are safe.

There was also a risk of double free if an exception occured in __init__
after data, col_index and row_index were set, but before _scipy was.  In
this case, the __dealloc__ routine would see that there was no _scipy,
so would deallocate pointers that were owned by NumPy.  When NumPy then
deallocated its matrices, we got the double free.

We make this safer by adding a _deallocate property to CSR, which is set
to False when the data attributes are set, or when the ownership is
given to numpy.  We set this to True in __cinit__ because we want the
default constructors to be factory methods which allocate C pointers and
store them, and we want a fail-safe mechanism for preventing memory
leaks.  As long as the only way to initialise directly in Python-space
code is __init__, we just take extra care within that to get better
memory handling elsewhere in the C code.