Remove useless test from test_basis_transformation
This didn't actually test what it purported to test - it was actually a
very round-about way of testing that `Qobj.eigenstates` produced the
same output of `sparse.sp_eigs`, which it of course does, because the
former just calls the latter.

The test could have had use in tesing `Qobj.eigenstates`, but that's
already tested elsewhere.