Add mathematical class methods to CSR type
The original plan had specifically been _not_ to add these, and to use
the dispatch operations.  Really, however, having to completely avoid
using infix operator notation in deep, complex code just makes it
unreadable.  We still don't want to bind in the dispatchers as the
special mathematical methods, because this will end up with quite a lot
of inter-dependent spaghetti in package organisation, and disallows
various possible situations where we may want to use context managers to
temporarily alter the dispatcher lookups---if they're stored all over
the place, it's much harder to manage.

As a compromise, we put in the mathematical methods as non-dispatched
operations, which only operate on CSR types.  All other types must mass
through the correct dispatcher.  Deep mathematical code is expected to
homogenise its inputs on entry, so that this non-dispatch requirement
should not be an issue.