Add eigensolvers for CSR
This is nearly a direct copy from the old `sparse.py`.  It is not
implemented in Cython because there is no need---all the heavy lifting
is done within scipy anyway, so there's nothing to be gained.

The structure of this could do with a clean-up, but the intricacies of
sparse eigenvalue code are tricky, especially when handling degeneracies
and eigenvalues close to 0, so it's quite a large amount of work for
very little benefit.