Add permute for CSR
This is a complete re-implementation of `core.sparse.sp_permute` in
Cython, but it guarantees that the output indices are sorted, and always
checks the permutation is valid.  This input check takes a negligible
amount of time (~1-2µs or so), and the sorting for very large, dense
matrices can contribute about 50% of the time spent, but in almost all
cases this new method is much faster than the previous version.  In
particular, we do not go via the COO format which saves some more Cython
code and some memory allocations.

At the moment, there are two methods of permutation used in the code -
`core.sparse.sp_permute` and `core.permute._permute`, even though at the
lowest level they're doing the same thing.  `permute` is moved to the
data layer, and `Qobj.permute` will change to use it---the new
`Qobj.permute` will simply convert a `dims` permutation into a `shape`
permutation, and pass off the heavy lifting to the data layer.