Add norm module for CSR
This allows replacing the rest of the `core.sparse` module.

In the backing Cython we always use BLAS functions where available since
they may well be hardware-specific, especially if MKL is available.
When AVX is available this can be good for a 2x speed increase on x86_64
systems.

Beware that the BLAS function `?asum` and `i?amax` are not generally
suitable for our use cases here---they work on `abs(z.real)+abs(z.imag)`
rather than the complex absolute value `abs(z)`.