Add permute by both indices and dimensions
The former is used in `steadystate` when you want to change the
bandwidth of a matrix that you have, and you want to move around every
single row and column to somewhere else.

The latter is the more common quantum operation, where we may want to
modify the tensor structure of a quantum object after the fact.  At the
moment the dimension handling will be done somewhat messily elsewhere,
but improving this likely needs an overhaul of the whole
dimension-handling code.

This is a major improvement over the old permutation code, as it
requires no temporary structures and conversions, and requires far fewer
computations for dense matrices.  For dense matrices, we first get the
whole row and column permutation, then do it by an index method like
mentioned earlier.  If the matrix is very sparse, however, we only
calculate new indices for the elements that we will actually use - the
logic is handled internally by the `_Indexer` class.  This is mostly
just a convenience to handle management of the internal state needed to
index.  There are a few pointers that need to be allocated and freed at
suitable times, and this just simplifies the process at a small speed
cost.

We're still an awful lot faster than the old code, however.