Make CSR sorting available externally
This defines a huge `Sorter` class for the CSR type which exposes the
mechanism to sort the indices.  This exposes two methods: one to sort a
created array in-place, and one to translate indices from an unsorted
location into a sorted one.  This is particularly useful when dealing
with mathematical operations which may produce unsorted indices on a
small scale.

We will often be dealing with very sparse matrices, with many quantum
operators being up to tri-diagonal.  In order to minimise allocations
and data copying, the Sorter class has optimised fast-paths to handle
sorts with three or fewer elements per row.  These sorts minimise the
average number of comparisons and data writes that are made.

To sort "in-place" with more than three elements, we realistically have
to copy the data to somewhere useful, sort it there, and then memcpy it
back to where we want it.  This is because we have to sort two separated
arrays at the same time, and even if we know the order at the end,
deciding on the order in which to swap the elements is tricky, unless we
use swap space to do it.  If we're using swap space, we may as well just
copy the elements in, sort them there, and copy them back.

To sort out-of-place, we use a pointer-based argsort to avoid copying
data unnecessarily.  We want the same comparison function to always be
reusable, so we make a comparison function based on pointer-dereference,
actually sorting an array of pointers into the relevant arrays.  With
this done, we can simply cycle through, copying the result into the
target finding the offset into the arrays by pointer arithmetic.

When copying out-of-place we allow the `data` and `col_index` pointers
to be passed separately.  This is expecially useful for functions which
modify the columns within each row into an unsorted order.  These can
then be written into temporary space, and sorted on.