Fix Python arithmetic with CSR
Cython earlier than 3.0 follows the C-API for arithmetic operations on
CPython extension classes, not Python semantics.  This means that for
methods such as `__add__`, the operands `a + b` will always appear as
`__add__(a, b)`, regardless of whether this method is being tried on the
type of `a` or the type of `b`.  This is different to Python semantics,
which would do `a.__add__(b)`, and if that fails try `b.__radd__(a)`.

We have to be careful to handle these cases sensibly.