Make CSR pickleable
It is more convenient in Cython to define `__reduce__` then
`__getstate__` and `__setstate__` since we're typically working much
lower-level anyway.

For simplicity, we pickle ourselves as our scipy view, basically passing
all the hard work off onto scipy and numpy.  This most easily ensures
that all our data will be pickled safely, and be completely thread-safe
when transferring it.

The new function `fast_from_scipy` can be used to reconstruct a CSR type
from a scipy array, but it does no type checking at all.  It simply
wraps the arrays it finds for speed - this is intended as an internal
function which is only called on trusted inputs.