Make Dense picklable, constructable from pointer
Dense pickles by converting itself to a numpy array, much like CSR does
with scipy matrices.

As a second convenience, we provide a way for C code to wrap a raw
pointer in a Dense type with minimal overhead as well.  This
necessitates a more robust way of handling the deallocation, since we no
longer only want to deallocate if the numpy array exists.  This is only
intended for short-lived Dense instances, since otherwise it may be hard
to manage the memory correctly.