Add expect for CSR
CQobjEvo uses `expect` pretty heavily, so this is effectively providing
a fast-path for it, which will also be used by `qutip.expect`.  Unlike
the naive method using `inner_op`, we do not need to allocate temporary
storage for the adjoint.

`expect` automatically performs `<x|A|x>` if given a state, or `tr(A.B)`
if given an operator.  We also provide `expect_super` why performs
`tr(A.B)` where `B` is given in column-stacked format, the same as
`cy_expect_rho_vec`.