Convert Qobj to use the new CSR type
BREAKING CHANGES:
    - Qobj.data is no longer a scipy.sparse.csr_matrix
    - matrix multiplication of bra/ket produces a scalar
    - otherwise, matrix multiplication no longer contracts 1D spaces

This is a "temporary" conversion of sorts - Qobj uses the new data-layer
type CSR, but it isn't properly dispatched yet, so it still assumes that
it has CSR-formatted data.

Qobj is properly responsible for managing its dimensions.  Shape of the
underlying data buffers is handled at the data-layer, but it's important
that Qobj properly ensures that all dimensions are always valid for the
operands.

As part of this, matrix multiplication has been stripped of its
"interesting" contraction of 1D tensor spaces, which generally caused
quite a nuisance in a lot of cases, and didn't help all that much in
superoperators where it was meant to help out.  We set matrix
multiplication of bra and ket to produce a scalar rather than a Qobj,
which hopefully will make some code a bit neater.

We make a lot of major internal changes to how Qobj initialisation and
internal Qobj objects are created.  We massively speed up average Qobj
initialisation time by ensuring that we always pass in the type if we
know it (inferring it from the dimensions is a huge hit in Python
space), and constructing the new dimensions.  We also encertain that
data we specifically constructed for a Qobj instance is not copied,
since this is just a waste of time.

Overall, this can save up to ~100µs on some larger Qobj initialisations,
and in general helps keep all mathematical operations down.  This should
really help in situations which deal with small quantum objects like
qubits, and perform many mathematical operations on them.

We unify the tidyup code to ensure it's called every time it's sensible,
and it's centrally controlled.