Update test_ptrace.py to new API
We get rid of a dependence on the old `legacy.ptrace` and put in a
manual dense form of the partial trace by repeated application of
numpy's Einstein summation convention.  This should (hopefully) be
fairly clearly correct, so is suitable for a test, and it's much less
fragile than the old legacy Cython code (and let's us get rid of it!).