Make index iteration more explicit
Cython will compile the new and the old versions of this change to
equivalent code, but this is perhaps a little clearer.  In C, the for
loop `for (int x = 0; x < y; x += stride)` would not work with how we
modify `x` within the loop body, but Cython would always use temporary
variables to be safe.  This just makes it abundantly clear that we're
not intending to screw with the loop.