Add matmul methods for Dense
We translate _everything_ into a simple call to BLAS zgemm, so we can
take best advantage of pipelining, parallelism, fused-add-multiply and
other vectorised CPU instructions - we assume that whatever version of
BLAS we can link against will be faster at all this stuff than we could
ever hope to be.

BLAS level 3 only guarantees that zgemm will be available for
Fortran-ordered matrix inputs and outputs, but our Dense type can be
either C- or Fortran-ordered.  MKL provides the switch CblasOrder, but
we don't need that if we're prepared to get our hands dirty, and we
can't guarantee that we'll link against MKL anyway, so we make use
of the fact that
    A . B == (B' . A')'
and that re-interpreting a Fortran-ordered matrix A as C-ordered is
equivalent to implicitly taking A' _without memory copy_.

Now we just have to decide: if the output is C-ordered, then we put
B'.A' (i.e. (A.B)') into it in Fortran-order.  If it's Fortran-ordered,
we put in A.B like normal.

Now for each of A and B, if we have them in C order, we treat it as if
we actually have A' or B'.  If we want A, we tell zgemm to transpose it,
and so on.  If we have them in Fortran order, we only transpose them if
we _actually_ want to operate on A'.

This means that we can perform the entire operation without allocating
any memory, no matter what the inputs A and B are.  This is at least as
fast as `np.dot` for the simplest case of `A . B` - we are faster at
very small sizes because we can elide a lot of the type-checking that
they have to do, e.g. we know our strides are 1.  The advantages are
most pronounced at medium-sized compound operations `a*A.B + C`, which
numpy cannot express in a single line of code, so requires additional
allocations.  At large sizes, the difference will shrink again, because
matrix multiplication scales (approximately) as cubic in a
characteristic dimension, while the allocations, scalar multiplications
and additions are only quadratic.