Use Python memory management in CSR sort
I am not certain this is necessarily helpful, in particular because it
requires acquiring the GIL to do the allocations.  For small chunks of
data in sorting, the Python memory manager will allocate into stack
memory, or other small-scale managed memory, and so should be faster
than a full call to malloc.  By acquiring the GIL, we know we're safe
with respect to parallelisation though I'm not sure it's actually a
problem at the moment, just I have occasional segfaults I'm struggling
to pin down.