Use _fix_flags whenever returning an ndarray
Various in-place operations on Dense matrices can change the shape,
which should be reflected in the numpy array we output.  Since we
guarantee that the result of `self.as_ndarray()` is always the same
object, the changes we make here will also propagate through to that
output (though not to derived views onto it, but there's not much we can
do about that).