Use new data-layer types in mesolve
This still currently makes assumptions about the data-layer types, but
it's easy to fix.  No measurable change in the run-time of tests.

The data-layer types _always_ output a 2D matrix when returning values,
which is a minor internal change.  Since the second dimension is
_always_ one, there's no real problem in almost all cases, except trying
to do operations with raw numpy arrays compared to the output of the
data layer types will sometimes cause weird quirks because the 1D array
will be broadcast up to match the 2D one.

Once _everything_ uses the data-layer types properly, this won't be an
issue anymore.