Make mcsolve use new data-layer types
This turned out to be much easier than expected.  It was sufficiently
easy that I suspect the conversion may not have been complete, but the
tests are not completely comprehensive.

Also removed some code which wasn't used, and would have caused errors
had it actually been called: the `_func_set` output took additional
integrator arguments, but the way the code actually called these, it
should have been a single tuple.  Also, in the `set_args` functions, the
name `psi0` was not defined, and would have thrown an error had it been
run.