Make steadystate work with new data-layer types
This is mostly a total fake-out.  We actually always use scipy matrices
underneath, because it has too many bits and bobs which are built on
them.  A lot of Cython code is deprecated in favour of the scipy
versions (which originally came from QuTiP!), or moved purely into the
Cython backing `_steadystate.pyx` for this module.

A couple of naming bugs are (hopefully) fixed - those paths clearly
aren't being tested, because there were undefined variables and it
couldn't work.