Fix possible divide-by-zero errors in permute.pyx
The possible error in `Indexer` was hugely unlikely to happen (the
passed dimensions would need to have a 0 in the `Qobj.dims` parameter),
but the error in dispatching based on the sparsity of a CSR type would
happen if `Qobj.permute` was called on a zero matrix.  Physically this
is useless, but an entirely valid input nonetheless.